Extract to file and rename jumpTracker
